[ru_egrul] Ingest all archives to build historic Ownership/Directoship
For every day, we get new archive files containing a full copy of the
changed companies. What we do is to build database and then, for each
updated company, fuse the new records with the previous records write
the merged record to the database. Right now, that's just creating
historic Ownership/Directorship entities.